version 1.0
NOTE: 1.0 here doesn't mean anything other than a major version change,
because the layout changed drastically with respect to previous
versions.

- fully incorporated home row mods on the Base layer

The mod distribution is different on each half for Alt and OS (GUI):

On the left, there's S/Alt, D/Control, F/Shift, and GUI is mapped to a
superkey that includes "Magic Sysrq" behavior. It can be used both as
standard GUI key, as a shortcut to Alt-Sysrq or as a shortcut (via
macro) to my version of the famous REISUB.

On the right the Alt is shifted outwards to pinky, because I use AltGr
for special characters in ES/PT/DE, and if I leave it on the ring finger
it makes for weird contortions when using it with the letter O. Also,
the right GUI is just a dual with the \| key, instead of the Superkey.

I also kept the dual Alts on the thumb space keys, and added Ctrl and
Shift on the left thumb cluster for convenience, to use it as an
alternative with the z/x/c/v/ins/del combos, and other combos such as
those from my WM.

- other Base changes

Added PrintScr which can be invoked with Shift/Alt etc for different
screenshot modes in my OS

Added a dedicated Capslock, the next step is to turn it into a Superkey
with optional auto capitalize word, and maybe auto capitalize sentence.

- restructured Lower

Added a second row of function keys, so it can be used for all sorts of
software/OS shortcuts.

On lower, the home row mods are instead fixed to the modifiers, and
since there's no need to combine (or use) AltGr, I just grouped the
Shift/Ctrl/Alts at the first three home positions.

Redistributed the media keys on the thumb clusters, and made both T8s to
OSL to Lower. These two OSL are used for A LOT:

* one handed media controls
* Function key stuff combined with HMRs (for example now I can have
  easier access for free to F11 and F12, those are fullscreen/maximize
  terminal and call Quake-style terminal whenever I am for example)
* jump to other layers, chording from the Layer navigation in Lower

While I'm super excited about the layout and it work better than any
other version before, it's not without bugs, particularly when it comes
to Superkey behaviour. For a feature so central to how we program this
keyboard, its implementation is poor, buggy and incomplete. So besides
counting to 100 and inspiring/expiring, if you can test this layout and
help me to improve the Superkeys I would be so grateful!