version 1.2
- added an F24 in Lower at the B position, that I map in my OS as a
  shortcut to run Bazecor

- added layers, macros, superkeys and bindings for the following kinds
  of writing helpers: Punctuation, Aphostrophe, Caps Lock modes. All can
  be used to assist when typing in L1 (Base).

Punctuation helpers:

Hold LT3 and press comma, colon/semicolon, dot or dash to add an extra
space after the punctuation character. Dot also capitalizes the first
letter after the space.

Aphostrophe helpers:

Hold LT3 and type ' (aphostrophe), i(I/it), y(you), s(she), h(he),
w(we), t(they), then hit l('ll), d('d), r('re), m('m) to complete the
aphostrophe form and add a space after.

This honors shift, so you can use it capitalizing the first letter ("I"
is always capitalized), or not.

To differentiate between I and it, press once or twice i.

Caps Lock modes

Caps lock is a Superkey that on tap behaves as Caps Lock and stays
normally in Base. A second keypress presses Caps Lock and enters a layer
with the following behaviours:

comma, dot, colon/semicolon, space, enter, tab and esc all behave as in
Lower for the first keypress, on second keypress they exit the Caps Lock
layer.

This adds a bit more fluidity and simulates "caps word", "caps
sentence", and "caps paragraph".

---

There is room for improvements so if anyone is interested in checking
these ideas it'd be cool. Also note that this has some idiosyncracies
that obviously stem from general compatibility with the whole layout,
there are different ways to enter/exit modes that might work better for
other layouts.

I get the feeling that there should be a way to combine these helpers on
the fly (for example using punctuation helpers while on extended
capslock mode), but I still couldn't quite nail that. Not sure if it
would need support from Bazecor there.

As to usefulness, I haven't yet used something like this for real
typing, so I'll give it a go and see if it brings improvements to the
table. I did this mostly out of curiosity about whether this kind of
stuff could be done with Bazecor. Conclusion, it is possible, but
involves a degree of pain, limitations, finding new bugs. But it's
possible and that's good